GridapROMs.Distributed

GridapROMs.Distributed.GenericPArray — Type

struct GenericPArray{V,A,B,C,D,T,N} <: AbstractArray{T,N}
  array_partition::A
  index_partition::B
  unpartitioned_axes::C
  cache::D
end

Same as PVector, but while the latter always stores a vector with entries partitioned on different cores, this structure stores an array (not necessarily a vector) partitioned along the first dimension (row-wise)

GridapROMs.Distributed.GenericPArray — Method

GenericPArray(array_partition,index_partition)

Create an instance of GenericPArray from the underlying properties array_partition and index_partition.

GridapROMs.Distributed.OwnAndGhostParamVectors — Type

struct OwnAndGhostParamVectors{A,B,T} <: AbstractParamVector{T}
  own_values::A
  ghost_values::A
  permutation::B
end

Parametric version of OwnAndGhostVectors