GridapROMs.ParamFESpaces

GridapROMs.ParamGeometry.DisplacementMap — Type

struct DisplacementMap <: GridMapStyle end

Trait for displacement maps d: Ω ↦ (I+d)∘Ω, with I being the identity; here, each coordinate x of Ω is mapped to x + d(x). Note: d may be parameterized (see AbstractParamFunction)

GridapROMs.ParamGeometry.GridMapStyle — Type

abstract type GridMapStyle end

Abstraction for maps used to modify a Grid in Gridap. Subtypes:

PhysicalMap

DisplacementMap

GridapROMs.ParamGeometry.ParamGrid — Type

abstract type ParamGrid{Dc,Dp} <: Grid{Dc,Dp} end

Abstraction for mapped grids, whenever the map is parameterized. Subtypes:

ParamMappedGrid

ParamUnstructuredGrid

GridapROMs.ParamGeometry.ParamMappedDiscreteModel — Type

struct ParamMappedDiscreteModel{Dc,Dp} <: DiscreteModel{Dc,Dp}
  model::DiscreteModel{Dc,Dp}
  mapped_grid::ParamMappedGrid{Dc,Dp}
end

Represents a model with a ParamMappedGrid grid. See also MappedDiscreteModel in Gridap.

GridapROMs.ParamGeometry.ParamMappedGrid — Type

struct ParamMappedGrid{Dc,Dp,A} <: ParamGrid{Dc,Dp}
  grid::Grid{Dc,Dp}
  node_coords::A
end

Standard implementation of a ParamGrid. The field grid represents a Grid in a reference configuration, while node_coords is a structure collecting the mapped coordinates

GridapROMs.ParamGeometry.ParamSubCellTriangulation — Type

struct ParamSubCellTriangulation{Dc,Dp,T,A} <: Triangulation{Dc,Dp}
  subcells::ParamSubCellData{Dc,Dp,T}
  bgmodel::A
  subgrid::ParamUnstructuredGrid{Dc,Dp,T,NonOriented,Nothing}
end

Parameterized version of a SubCellTriangulation in GridapEmbedded

GridapROMs.ParamGeometry.ParamSubFacetTriangulation — Type

struct ParamSubFacetTriangulation{Dc,Dp,T,A} <: Triangulation{Dc,Dp}
  subfacets::ParamSubFacetData{Dp,T}
  bgmodel::A
  subgrid::ParamUnstructuredGrid{Dc,Dp,T,NonOriented,Nothing}
end

Parameterized version of a SubFacetTriangulation in GridapEmbedded

GridapROMs.ParamGeometry.ParamUnstructuredGrid — Type

struct ParamUnstructuredGrid{Dc,Dp,Tp,O,Tn} <: ParamGrid{Dc,Dp}
  node_coordinates::ParamBlock{Vector{Point{Dp,Tp}}}
  cell_node_ids::Table{Int32,Vector{Int32},Vector{Int32}}
  reffes::Vector{LagrangianRefFE{Dc}}
  cell_types::Vector{Int8}
  orientation_style::O
  facet_normal::Tn
  cell_map
end

Parameterized version of a Gridap UnstructuredGrid. Note: this struct could be avoided in future versions, by simply generalizing the type of the node_coordinates field in UnstructuredGrids.

GridapROMs.ParamGeometry.PhysicalMap — Type

struct PhysicalMap <: GridMapStyle end

Trait for standard physical maps φ: Ω ↦ φ∘Ω, i.e. each coordinate x of Ω is mapped to φ(x). Note: φ may be parameterized (see AbstractParamFunction)

GridapROMs.ParamGeometry.mapped_grid — Method

mapped_grid(style::GridMapStyle,grid::Grid,f) -> Grid

Returns a Grid by applying the map f on a reference configuration grid, according to the strategy provided by style