GridapROMs.ParamFESpaces

GridapROMs.ParamFESpaces.MultiFieldParamFEFunction — Type

struct MultiFieldParamFEFunction{T<:MultiFieldCellField} <: ParamFEFunction
  single_fe_functions::Vector{<:SingleFieldParamFEFunction}
  free_values::AbstractArray
  fe_space::MultiFieldFESpace
  multi_cell_field::T
end

GridapROMs.ParamFESpaces.MultiFieldParamFESpace — Type

const MultiFieldParamFESpace = MultiFieldFESpace

GridapROMs.ParamFESpaces.ParamFEFunction — Type

abstract type ParamFEFunction <: FEFunction end

Parametric extension of a FEFunction in Gridap. Subtypes:

SingleFieldParamFEFunction

MultiFieldParamFEFunction

GridapROMs.ParamFESpaces.ParamTrialFESpace — Type

const ParamTrialFESpace = UnEvalTrialFESpace

GridapROMs.ParamFESpaces.SingleFieldParamFEFunction — Type

struct SingleFieldParamFEFunction{T<:CellField} <: ParamFEFunction
  cell_field::T
  cell_dof_values::AbstractArray{<:ParamBlock}
  free_values::AbstractParamVector{<:Number}
  dirichlet_values::AbstractParamVector{<:Number}
  fe_space::SingleFieldFESpace
end

GridapROMs.ParamFESpaces.SingleFieldParamFESpace — Type

abstract type SingleFieldParamFESpace{S} <: SingleFieldFESpace end

Parametric extension of a SingleFieldFESpace in Gridap. The FE spaces inhereting are (trial) spaces on which we can easily define a ParamFEFunction.

Subtypes:

TrivialParamFESpace

TrialParamFESpace

GridapROMs.ParamFESpaces.TrialParamFESpace — Type

TrialParamFESpace{S} <: SingleFieldParamFESpace{S}

Trial FE space equipped with parametric dirichlet values

GridapROMs.ParamFESpaces.TrivialParamFESpace — Type

TrivialParamFESpace{S} <: SingleFieldParamFESpace{S}

Wrapper for non-parametric FE spaces that we wish assumed a parametric length

GridapROMs.ParamFESpaces.UnEvalTrialFESpace — Type

struct UnEvalTrialFESpace{A,B} <: SingleFieldFESpace
  space::A
  space0::B
  dirichlet::Union{Function,AbstractVector{<:Function}}
end

Struct representing trial FE spaces that are not evaluated yet. This may include FE spaces representing transient problems (although the implementation in Gridap, called TransientTrialFESpace, does not fall into this category), parametric problems, and transient-parametric problems.

GridapROMs.ParamDataStructures.parameterize — Method

parameterize(a::SparseMatrixAssembler,plength::Int) -> SparseMatrixAssembler

Returns an assembler that also stores the parametric length of r. This function is to be used to assemble parametric residuals and Jacobians. The assembly routines follow the same pipeline as in Gridap

GridapROMs.ParamFESpaces.HomogeneousTrialParamFESpace — Method

HomogeneousTrialParamFESpace(U::SingleFieldFESpace,plength::Int) -> TrialParamFESpace

Returns a TrialParamFESpace equipped with homogeneous parametric dirichlet values

GridapROMs.ParamFESpaces.TrialParamFESpace! — Method

TrialParamFESpace!(dir_values::AbstractParamVector,space::SingleFieldFESpace,objects
  ) -> TrialParamFESpace

Allows do-block syntax for the construction of a TrialParamFESpace

GridapROMs.ParamFESpaces.get_vector_type2 — Method

get_vector_type2(f::SingleFieldParamFESpace) -> Type

Returns the vector type of the underlying un-parametric FESpace contained in f

GridapROMs.ParamFESpaces.param_zero_dirichlet_values — Function

param_zero_dirichlet_values(f::FESpace,L::Integer=param_length(f)) -> AbstractParamVector

Parametric version of the function zero_dirichlet_values in Gridap

GridapROMs.ParamFESpaces.param_zero_free_values — Function

param_zero_free_values(f::FESpace,L::Integer=param_length(f)) -> AbstractParamVector

Parametric version of the function zero_free_values in Gridap