GridapROMs.ParamODEs

GridapROMs.ParamODEs.JointODEParamOperator — Type

const JointODEParamOperator{O<:ODEParamOperatorType} = ODEParamOperator{O,JointDomains}

GridapROMs.ParamODEs.JointTransientParamFEOperator — Type

const JointTransientParamFEOperator{O<:ODEParamOperatorType} = TransientParamFEOperator{O,JointDomains}

GridapROMs.ParamODEs.LinearNonlinearParamODE — Type

struct LinearNonlinearParamODE <: ODEParamOperatorType end

GridapROMs.ParamODEs.LinearParamODE — Type

struct LinearParamODE <: ODEParamOperatorType end

GridapROMs.ParamODEs.NonlinearParamODE — Type

struct NonlinearParamODE <: ODEParamOperatorType end

GridapROMs.ParamODEs.ODEParamOperator — Type

const ODEParamOperator{T<:ODEParamOperatorType,T<:TriangulationStyle} = ParamOperator{O,T}

Transient extension of the type ParamOperator.

GridapROMs.ParamODEs.ODEParamOperatorType — Type

abstract type ODEParamOperatorType <: UnEvalOperatorType end

Parametric extension of the type ODEOperatorType in Gridap.

Subtypes:

NonlinearParamODE

LinearParamODE

LinearNonlinearParamODE

GridapROMs.ParamODEs.ODEParamSolution — Type

struct ODEParamSolution{V} <: ODESolution
  solver::ODESolver
  odeop::ODEParamOperator
  r::TransientRealization
  us0::Tuple{Vararg{V}}
end

GridapROMs.ParamODEs.ParamStageOperator — Type

struct ParamStageOperator{O} <: NonlinearParamOperator
  op::ODEParamOperator{O}
  r::TransientRealization
  state_update::Function
  ws::Tuple{Vararg{Real}}
  paramcache::AbstractParamCache
end

Stage operator to solve a parametric ODE with a time marching scheme

GridapROMs.ParamODEs.SplitODEParamOperator — Type

const SplitODEParamOperator{O<:ODEParamOperatorType} = ODEParamOperator{O,SplitDomains}

GridapROMs.ParamODEs.SplitTransientParamFEOperator — Type

const SplitTransientParamFEOperator{O<:ODEParamOperatorType} = TransientParamFEOperator{O,SplitDomains}

GridapROMs.ParamODEs.TransientLinearParamFEOpFromWeakForm — Type

struct TransientLinearParamFEOpFromWeakForm{T} <: TransientParamFEOperator{LinearParamODE,T}
  res::Function
  jacs::Tuple{Vararg{Function}}
  constant_forms::Tuple{Vararg{Bool}}
  tpspace::TransientParamSpace
  assem::Assembler
  trial::FESpace
  test::FESpace
  domains::FEDomains
  order::Integer
end

Instance of TransientParamFEOperator, to be used when the transient problem is linear

GridapROMs.ParamODEs.TransientMultiFieldParamFESpace — Type

const TransientMultiFieldParamFESpace = MultiFieldFESpace

GridapROMs.ParamODEs.TransientParamFEOpFromWeakForm — Type

struct TransientParamFEOpFromWeakForm{T} <: TransientParamFEOperator{NonlinearParamODE,T}
  res::Function
  jacs::Tuple{Vararg{Function}}
  tpspace::TransientParamSpace
  assem::Assembler
  trial::FESpace
  test::FESpace
  domains::FEDomains
  order::Integer
end

Instance of TransientParamFEOperator, to be used when the transient problem is nonlinear

GridapROMs.ParamODEs.TransientParamFEOperator — Type

const TransientParamFEOperator{O<:ODEParamOperatorType,T<:TriangulationStyle} = ParamFEOperator{O,T}

Parametric extension of a TransientFEOperator in Gridap. Compared to a standard TransientFEOperator, there are the following novelties:

a TransientParamSpace is provided, so that parametric realizations can be extracted directly from the TransientParamFEOperator

a function representing a norm matrix is provided, so that errors in the desired norm can be automatically computed

Subtypes:

TransientParamFEOpFromWeakForm

TransientLinearParamFEOpFromWeakForm

GridapROMs.ParamODEs.TransientParamFESolution — Type

struct TransientParamFESolution{V} <: TransientFESolution
  odesol::ODEParamSolution{V}
  trial
end

Wrapper around a TransientParamFEOperator and ODESolver that represents the parametric solution at a set of time steps. It is an iterator that computes the solution at each time step in a lazy fashion when accessing the solution.

GridapROMs.ParamODEs.TransientTrialParamFESpace — Type

const TransientTrialParamFESpace = UnEvalTrialFESpace

GridapROMs.ParamODEs.TransientLinearParamFEOperator — Method

TransientLinearParamFEOperator(res::Function,forms::Tuple{Vararg{Function}}, tpspace,trial,test;kwargs...) -> TransientLinearParamFEOpFromWeakForm{TriangulationStyle}

Returns a linear parametric FE operator