GridapROMs.ParamSteady

GridapROMs.ParamSteady.GenericParamOperator — Type

struct GenericParamOperator{O<:UnEvalOperatorType,T<:TriangulationStyle} <: ParamOperator{O,T}
  op::ParamFEOperator{O,T}
end

Wrapper that transforms a ParamFEOperator into an ParamOperator

GridapROMs.ParamSteady.JointParamFEOperator — Type

const JointParamFEOperator{O<:UnEvalOperatorType} = ParamFEOperator{O,JointDomains}

GridapROMs.ParamSteady.JointParamOperator — Type

const JointParamOperator{O<:UnEvalOperatorType} = ParamOperator{O,JointDomains}

GridapROMs.ParamSteady.LinearNonlinearParamEq — Type

struct LinearNonlinearParamEq <: UnEvalOperatorType end

GridapROMs.ParamSteady.LinearNonlinearParamFEOperator — Type

struct LinearNonlinearParamFEOperator{O<:UnEvalOperatorType,T<:TriangulationStyle} <: ParamFEOperator{O,T}
  op_linear::ParamFEOperator
  op_nonlinear::ParamFEOperator
end

Interface to accommodate the separation of terms depending on their linearity in a nonlinear problem. This allows to build and store once and for all linear residuals/Jacobians, and in the Newton-like iterations only evaluate and assemble only the nonlinear components

GridapROMs.ParamSteady.LinearParamEq — Type

struct LinearParamEq <: UnEvalOperatorType end

GridapROMs.ParamSteady.NonlinearParamEq — Type

struct NonlinearParamEq <: UnEvalOperatorType end

GridapROMs.ParamSteady.ParamFEOpFromWeakForm — Type

struct ParamFEOpFromWeakForm{O<:UnEvalOperatorType,T<:TriangulationStyle} <: ParamFEOperator{O,T}
  res::Function
  jac::Function
  pspace::ParamSpace
  assem::Assembler
  trial::FESpace
  test::FESpace
  domains::FEDomains
end

Most standard instance of a parametric FE operator

GridapROMs.ParamSteady.ParamFEOperator — Type

abstract type ParamFEOperator{O<:UnEvalOperatorType,T<:TriangulationStyle} <: FEOperator end

Parametric extension of a FEOperator in Gridap. Compared to a standard FEOperator, there are the following novelties:

a ParamSpace is provided, so that parametric realizations can be extracted directly from the ParamFEOperator

a function representing a norm matrix is provided, so that errors in the desired norm can be automatically computed

Subtypes:

ParamFEOpFromWeakForm

LinearNonlinearParamFEOperator

TransientParamFEOperator

GridapROMs.ParamSteady.ParamOperator — Type

abstract type ParamOperator{O<:UnEvalOperatorType,T<:TriangulationStyle} <: NonlinearParamOperator end

Type representing algebraic operators when solving parametric differential problems

GridapROMs.ParamSteady.SplitParamFEOperator — Type

const SplitParamFEOperator{O<:UnEvalOperatorType} = ParamFEOperator{O,SplitDomains}

GridapROMs.ParamSteady.SplitParamOperator — Type

const SplitParamOperator{O<:UnEvalOperatorType} = ParamOperator{O,SplitDomains}

GridapROMs.ParamSteady.UnEvalOperatorType — Type

abstract type UnEvalOperatorType <: GridapType end

Type representing operators that are not evaluated yet. This may include operators representing transient problems (although the implementation in Gridap differs), parametric problems, and a combination thereof. Could become a supertype of ODEOperatorType in Gridap

GridapROMs.ParamSteady.LinearParamFEOperator — Method

LinearParamFEOperator(res::Function,jac::Function,pspace,trial,test,args...
  ) -> ParamFEOperator{LinearParamEq,TriangulationStyle}

Returns a linear parametric FE operator

GridapROMs.ParamSteady.get_dof_map_at_domains — Method

get_dof_map_at_domains(op::ParamFEOperator) -> Contribution

Returns the residual dof map restricted to every residual triangulation

GridapROMs.ParamSteady.get_fe_operator — Method

get_fe_operator(op::ParamOperator) -> ParamFEOperator

Fetches the underlying FE operator of an algebraic operator op

GridapROMs.ParamSteady.get_param_space — Method

get_param_space(feop::ParamFEOperator) -> ParamSpace
get_param_space(feop::TransientParamFEOperator) -> TransientParamSpace

Returns the space of parameters contained in feop

GridapROMs.ParamSteady.get_sparse_dof_map_at_domains — Method

get_sparse_dof_map_at_domains(op::ParamFEOperator) -> Contribution

Returns the Jacobian dof map restricted to every Jacobian triangulation

GridapROMs.ParamSteady.join_operators — Method

join_operators(op::LinearNonlinearParamFEOperator) -> ParamFEOperator

Joins the linear/nonlinear parts of the operator and returns the resulting operator

GridapROMs.Utils.change_domains — Method

change_domains(op::ParamFEOperator,args...) -> ParamFEOperator

Changes the triangulations for residual/jacobian; the resulting operator will have the same TriangulationStyle trait as the one of the input

GridapROMs.Utils.set_domains — Method

set_domains(op::ParamFEOperator,args...) -> JointParamFEOperator

Fixes the triangulations for residual/jacobian; the resulting operator will have the trait JointDomains activated