GridapROMs.Utils

GridapROMs.Utils.ArrayContribution — Type

const ArrayContribution{T,N} = Contribution{<:Union{AbstractArray{T,N},ArrayBlock{T,N}}}

Contribution whose field values are AbstractArrays

GridapROMs.Utils.Contribution — Type

struct Contribution{V}
  values::Tuple{Vararg{V}}
  trians::Tuple{Vararg{Triangulation}}
end

Collection of values corresponding to a set of triangulations. Similarly to DomainContribution, the values can be accessed by indexing the corresponding triangulation.

GridapROMs.Utils.DomainOperator — Type

GridapROMs.Utils.FEDomainOperator — Type

GridapROMs.Utils.FEDomains — Type

struct FEDomains{A,B}
  domains_res::A
  domains_jac::B
end

Fields:

domains_res: triangulations relative to the residual (nothing by default)

domains_jac: triangulations relative to the Jacobian (nothing by default)

GridapROMs.Utils.JointDomains — Type

struct JointDomains <: TriangulationStyle end

Trait for a FE operator indicating that residuals/Jacobians in this operator should be computed summing the contributions relative to each triangulation as occurs in Gridap

GridapROMs.Utils.JointFEOperator — Type

GridapROMs.Utils.JointOperator — Type

GridapROMs.Utils.LinearEq — Type

GridapROMs.Utils.LinearNonlinearEq — Type

GridapROMs.Utils.MatrixContribution — Type

const MatrixContribution{T} = ArrayContribution{T,2}

GridapROMs.Utils.NonlinearEq — Type

GridapROMs.Utils.OperatorType — Type

GridapROMs.Utils.PartialDerivative — Type

struct PartialDerivative{N} <: Function end

Gridap Map that implements a partial derivative

GridapROMs.Utils.SplitDomains — Type

struct SplitDomains <: TriangulationStyle end

Trait for a FE operator indicating that residuals/Jacobians in this operator should be computed keeping the contributions relative to each triangulation separate

GridapROMs.Utils.SplitFEOperator — Type

GridapROMs.Utils.SplitOperator — Type

GridapROMs.Utils.TriangulationStyle — Type

abstract type TriangulationStyle <: GridapType end

Subtypes:

JointDomains

SplitDomains

GridapROMs.Utils.TupOfArrayContribution — Type

const TupOfArrayContribution{T} = Tuple{Vararg{ArrayContribution{T}}}

Specifically allows to deal with tuples of Jacobians in unsteady settings

GridapROMs.Utils.VectorContribution — Type

const VectorContribution{T} = ArrayContribution{T,1}

GridapROMs.Utils.∂₁ — Type

const ∂₁ = PartialDerivative{1}

GridapROMs.Utils.∂₂ — Type

const ∂₂ = PartialDerivative{2}

GridapROMs.Utils.∂₃ — Type

const ∂₃ = PartialDerivative{3}

GridapROMs.Utils.collect_cell_matrix_for_trian — Method

function collect_cell_matrix_for_trian(
  trial::FESpace,
  test::FESpace,
  a::DomainContribution,
  strian::Triangulation
  ) -> Tuple{Vector{<:Any},Vector{<:Any},Vector{<:Any}}

Computes the cell-wise data needed to assemble a global sparse matrix for a given input triangulation strian

GridapROMs.Utils.collect_cell_vector_for_trian — Method

function collect_cell_vector_for_trian(
  test::FESpace,
  a::DomainContribution,
  strian::Triangulation
  ) -> Tuple{Vector{<:Any},Vector{<:Any}}

Computes the cell-wise data needed to assemble a global vector for a given input triangulation strian

GridapROMs.Utils.compute_error — Method

compute_error(
  sol::AbstractArray{T,N},
  sol_approx::AbstractArray{T,N},
  args...
  ) where {T,N} -> Number

Computes the error between sol and sol_approx, by default in the Euclidean norm. A different norm (usually represented by a sparse matrix) can be provided as an argument.

GridapROMs.Utils.compute_relative_error — Method

compute_relative_error(
  sol::AbstractArray{T,N},
  sol_approx::AbstractArray{T,N},
  args...
  ) where {T,N} -> Number

Computes the relative error between sol and sol_approx, by default in the Euclidean norm. A different norm (usually represented by a sparse matrix) can be provided as an argument.

GridapROMs.Utils.compute_speedup — Method

compute_speedup(t1::CostTracker,t2::CostTracker) -> Speedup

Computes the speedup the tracker t2 achieves with respect to t1, in time and in memory footprint

GridapROMs.Utils.contribution — Method

contribution(f,trians) -> Contribution

Constructor of a Contribution that allows do-block syntax. f is a function such that

values[i] = f(trians[i]) for i...

This constructor first builds the tuple of values, then builds the Contribution object from values and trians

GridapROMs.Utils.get_fe_operator — Method

get_fe_operator(op::ParamOperator) -> ParamFEOperator

Fetches the underlying FE operator of an algebraic operator op

GridapROMs.Utils.get_parent — Method

get_parent(t::Triangulation) -> Triangulation

When t is a triangulation view, returns its parent; throws an error when t is not a triangulation view

GridapROMs.Utils.get_polynomial_order — Method

get_polynomial_order(f::FESpace) -> Integer

Retrieves the polynomial order of f

GridapROMs.Utils.get_polynomial_orders — Method

get_polynomial_orders(fs::FESpace) -> Integer

Retrieves the polynomial order of fs for every dimension

GridapROMs.Utils.is_parent — Method

is_parent(parent::Triangulation,child::Triangulation) -> Bool

Returns true if child is a triangulation view of parent, false otherwise

GridapROMs.Utils.isapprox_parent — Method

isapprox_parent(parent::Triangulation,child::Triangulation) -> Bool

Same as is_parent, but with a relaxed check (it could return true even when the objectid comparison fails)

GridapROMs.Utils.order_domains — Method

order_domains(
  parents::Tuple{Vararg{Triangulation}},
  children::Tuple{Vararg{Triangulation}}
  ) -> Tuple{Vararg{Triangulation}}

Orders the triangulation children in the same way as the triangulation parents